# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ElorAdmin API",
        "version": "2.0"
    },
    "paths": {
...
        "/api/cycles": {
            "get": {
                "tags": [
                    "Cycles"
                ],
                "summary": "Shows cycles",
                "operationId": "f951d8ce358a5e631d599e533a1a4500",
                "responses": {
                    "200": {
                        "description": "Shows all cycles."
                    },
                    "204": {
                        "description": "There are no cycles."
                    },
                    "default": {
                        "description": "Error has ocurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cycles"
                ],
                "summary": "Create a cycle",
                "operationId": "7bab19dfe654a8051390639bd09db3b9",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Role Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "department_id",
                        "in": "query",
                        "description": "Department ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "modules[]",
                        "in": "query",
                        "description": "Modules",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cycle created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cycle": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "default": {
                        "description": "Error has ocurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}